EXUI-253: Move s2s token generation into test that needs it
